requirements.txt
python-telegram-bot==20.7
openai
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
